% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_R.R
\name{fix_beta_large}
\alias{fix_beta_large}
\title{Fixes beta value to get correct R0}
\usage{
fix_beta_large(
  params,
  S0,
  I,
  R0,
  use_eig = FALSE,
  beta = NULL,
  symp_trans = NULL
)
}
\description{
Fixes beta value to get correct R0
}
