% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_param_sets}
\alias{run_param_sets}
\title{Function to run multiple param sets}
\usage{
run_param_sets(
  paramsets,
  L = 100,
  N_particles = 1,
  N_threads_internal = 1,
  N_threads_external = 1,
  silent = TRUE
)
}
\description{
Function to run multiple param sets
}
