% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_params}
\alias{run_params}
\title{run_params}
\usage{
run_params(
  params,
  L = 200,
  N_particles = 1,
  N_threads = 1,
  run_name = "run1",
  run_params = list(),
  silent = TRUE,
  estimate_Rt = FALSE,
  Rt_per_day = 1
)
}
\description{
Run the metapopulation model with a set of parameters and a given number of particles and threads.
The number of particles gives the number of samples and threads the number of cores used to sample
}
